% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AusCovid19-package.R
\docType{package}
\name{AusCovid19-package}
\alias{AusCovid19}
\alias{AusCovid19-package}
\title{AusCovid19: Download and Visualize Covid-19 Related Data in Australia}
\description{
This package facilitates the download of cumulative and daily counts
    of confirmed COVID-19 cases, deaths and recoveries in the states of Australia
    from the Johns Hopkins University CSSE. In addition, the package features a 
    shiny app to visualize the spread of COVID-19 in Australia.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/etc5523-2020/r-package-assessment-dilinie-seimon}
  \item Report bugs at \url{https://github.com/etc5523-2020/r-package-assessment-dilinie-seimon/issues}
}

}
\author{
\strong{Maintainer}: Dilinie Seimon \email{asei0003@student.monash.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
