% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download Australia COVID-19 data from JHU CSSE}
\usage{
download_data(data_type = "cumulative")
}
\arguments{
\item{type}{A string, "cumulative" or "daily" defining the type of data to be downloaded}
}
\value{
A dataframe of confirmed COVID-19 cases, deaths and recoveries in Australia
}
\description{
Download Australia COVID-19 data from JHU CSSE
}
\examples{
download_data(data_type = "cumulative")
download_data(data_type = "daily")

}
