% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_value_box.R
\name{generate_value_box}
\alias{generate_value_box}
\title{Generate value box after summing a vector}
\usage{
generate_value_box(counts_vector, text, icon_name, color)
}
\arguments{
\item{counts_vector}{A numeric vector, To be summed up in generating the value printed}

\item{text}{A String, The text printed on the value box}

\item{icon_name}{A String, The font-awesome icon name}

\item{color}{A String, The color of the box}
}
\value{
An html widget including the summed vector, text and icon
}
\description{
Generate value box after summing a vector
}
\examples{
generate_value_box(counts_vector=c(1, 5, 4, 9, 0), text="Confirmed Cases", icon="head-side-mask", color="green")

x <- c(37,42,75,29,46,85)
generate_value_box(counts_vector=x, text="Confirmed Cases", icon="head-side-mask", color="green")

}
