% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_case_count.R
\name{get_case_count}
\alias{get_case_count}
\title{Get case count at date}
\usage{
get_case_count(data_frame, date, case_type)
}
\arguments{
\item{data_frame}{A dataframe, COVID-19 counts associated with dates, use download_data() function for dataframe}

\item{date}{A string, The date at which the case numbers are required}

\item{case_type}{A string, The type of case numbers required - either "Confirmed", "Deaths" or "Recovered"}
}
\value{
A vector of COVID-19 cases of required type and date
}
\description{
Get case count at date
}
\examples{
get_case_count(data_frame = download_data(data_type = "daily"), date = "2020-10-01", case_type = "Confirmed")

df <- download_data(data_type = "cumulative")
get_case_count(data_frame = df, date = "2020-10-01", case_type = "Deaths")


}
