% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_daily_counts.R
\name{calculate_daily_counts}
\alias{calculate_daily_counts}
\title{Calculate daily COVID case counts from cumulative counts}
\usage{
calculate_daily_counts(cum_counts_df)
}
\arguments{
\item{type}{A dataframe of cumulative confirmed cases, deaths and recoveries in Australia. Can be downloaded using download_data(data_type = "cumulative")}
}
\value{
A dataframe of daily confirmed COVID-19 cases, deaths and recoveries in Australia
}
\description{
Calculate daily COVID case counts from cumulative counts
}
\examples{
calculate_daily_counts(cum_counts_df = download_data(data_type = "cumulative")

df <- download_data(data_type = "cumulative")
calculate_daily_counts(cum_counts_df = df)

}
