% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch shiny app visualizing the spread of COVID-19 across the states of Australia}
\usage{
launch_app()
}
\value{
A Shiny Application
}
\description{
Launch shiny app visualizing the spread of COVID-19 across the states of Australia
}
